Stop installing Tempest at system wide for stable branch
As added in notes for INSTALL_TEMPEST variable we need to set
this as False for stable branch so that devstack does not
install Tempest at system wide.

- https://github.com/openstack/devstack/blob/aa2821eb89ab9c8048509d15fe41215c163d2f50/lib/tempest#L61

This should be done at the time when we cut the stable branch but
we forgot to do that for ussuri and victoria.

Change-Id: I23c77f98c2e969d8046d5212b883e343c36cd1b1
(cherry picked from commit 9a8c06231240f05fe136a40e2fc14b6f46e10471)