[stable-only] Preinstall pbr for devstack-tools
Due to "Deprecation of non-SNI compatible clients" [1], when
devstack-tools is installed in early phase (in neutron-grenade job for
example) and pbr is not yet installed, the pip command fails with the
following error in xenial based environments:

File "/usr/lib/python3/dist-packages/setuptools/command/easy_install.py", line 657, in easy_install
  raise DistutilsError(msg)
distutils.errors.DistutilsError: Could not find suitable distribution for Requirement.parse('pbr>=1.8')

This is because setuptools' easy_install is used, which is non-SNI
compatible in xenial base image (and python 3.5). With this patch the
issue can be avoided by preinstalling pbr just before the installation
of devstack-tools.

[1] https://github.com/pypa/pypi-support/issues/978

Change-Id: I0edd4734df8ac7976d3090b75ca4c033b8ff0f2c