[CI] [stable-only] Drop testing on Fedora
The QA/DevStack team does not support testing on Fedora on stable
branches. The fedora-latest nodeset is meant to be used on
master branch only.
This change removes it on this stable branch to allow the nodeset
to progress with Fedora version on master.
In addition, the two Fedora jobs are removed to allow for nodeset
removal.

Cherry-Picked-From: c1928d8fb89c451e98e7da1cc7f884c44d282dfd
Change-Id: If952b54bdecf82be811b57f4a17a0c59db54a62a