[stable-only] Use queens-eol for horizon
This patch is needed as tempest-full-py3 zuul job started to fail on
stable/queens branch due to the following:
Horizon's stable/queens transitioned to End of Life and its
stable/queens branch was deleted. This fix uses the queens-eol tag to
be able to check out horizon's latest state from queens.

Change-Id: I845f53de2d29ab7642ebb04e8bd0ce6c09ff12e8