Remove the zuul pragma to match master and feature/r1
This xena branch should not provide configs to master and feature/r1.
Doing so causes devstack's pre-run playbook to run twice as we run both
master's and xena's pre-run playbooks. This then fails because the
configure-swap role is not idempotent.

We fix this by detaching xena from master here and let master run master
configs and xena runs xena configs.

Change-Id: I1b3a463211b65b6aaa3e0e127a1b32962ad6bf9f