Test stable/stein with tempest 24.0.0
For every stable branch, Tempest as a range of versions
to work[1] and we can pick either first supported
or last supported version to test the specific EM stable
branch (when Tempest master is in not used).

For stable/stein, first supported version is Tempest 20.0.0
and last supported verson is 26.0.0.

Till now we used the last supported version and it was working
fine until jsonschema latest release 4.0.0 has some incompatible
changes. Now we can say our upper constraints should protect us
from this (like it protect in Tempest master case). But thing
is little tricky here on whether we use the compatible constraints
with tempest 26.0.0. or not.

Compatible constraints for Tempest 26.0.0 is stable/victoria constraints
(the time tag 26.0.0 was cut) but we used stable/stein constraints with
tempest 26.0.0 in stable/stein[2].

We should always use the compatible constraints and tag combination.

We have two option here to solve:

1. Use Tempest 24.0.0 and compatible stable/stein constraints (this commit)
2. Use Tempest 26.0.0 and compatible stable/victoria constraints

I am going with the 1st approach as it is simple and 2nd approach need
modification in run-tempest roles etc. Reason to pick the 24.0.0 was because
it has the nova novnc test fixes.

To make 24.0.0 work, we need to revert the below commit also:

Revert "Remove tempest deprecated img_dir option"

This reverts commit 9984c6ba250a772fd6e15018004ba0289e9cc1e2.

Reason for revert: bug#1946403

Co-Authored-By: melanie witt <melwittt@gmail.com>

[1] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html
[2] https://opendev.org/openstack/devstack/src/branch/stable/stein/stackrc#L319-L320

Change-Id: I98ec0029666aa9bf0d66c1f083d81cd7a20597fb