Drop CentOS8 job and nodeset
CentOS8 is going to be EOL soon and opendev
is planning to drop the centOS8 image from CI
- http://lists.openstack.org/pipermail/openstack-discuss/2022-January/026621.html

Devstack stable/wallaby and onwards has replaced the
centOS testing with centOS Stream. For older stable branches
(<= stable/victoria) we can either replace it with centOS Steam
or do the removal of CentOS8 and adding CentOS Stream testing
in two steps 1. drop the centOS8 testing as it cannot
be supported in CI 2. add centOS Stream job if possible.

This commit propose the step 1 as required by opendev image removal.

[Stable only]

Change-Id: I36751569d92fbc5084b8308d423a75318ae7d406