Pin Tempest to 26.1.0 tag for stable/train testing
Stable train is in Extended maintenance state[1] and
we need to cap stable/train with older Tempest.

But because of oslo.utils version bump in Tempest 27.0.0
and 28.0.0 which is not compatible with stable/train
upper constraints so these versions do not work in stable/train

- https://zuul.opendev.org/t/openstack/build/1146c9ce9f474d0d82b99b48b61c531e

Let's use the Tempest 26.1.0 which is also compatible with stable/train
testing.

Related-Bug: #1955418
Change-Id: I376be5f5a4d693aeaaf74ca5285269f796103c15