Add python3.6 pip support
Since pip v22, python3.6 is not supported (the minimum version is
python3.7). This patch adds the reference for the pip3.6 URL to be
used instead of the default one.

NOTE: stable/train had its own method to provide support for py3x.
This patch uses it instead of pushing the master patch approach.

This patch is also fixing the py35 pip URL string that should be
https://bootstrap.pypa.io/pip/3.5/get-pip.py.

Conflicts:
  tools/install_pip.sh

Closes-Bug: #1959600
Change-Id: Iab2c391d5388461fe9e9037cee81884ce8032e72