Remove the zuul pragma to match master and feature/r1
This yoga branch should not provide configs to master and feature/r1.
Doing so causes devstack's pre-run playbook to run twice as we run both
master's and yoga's pre-run playbooks. This then fails because the
configure-swap role is not idempotent.

We fix this by detaching yoga from master here and let master run master
configs and yoga runs yoga configs.

Change-Id: I958763a9329bd274415599a665b1295f59514bee